Test: Speculative fix for race condition in testAidConflictResolution_walletRoleEnabledNfcEnabledPreFixAid_walletWins
The test seems to have a race condition that only manifests in post-submit setups. Multiple attempts to reproduce this on local machines have failed, and the speculation is that its because the difference in CPU power. The test-case fails 100% when run on a 6.6 kernel, vs 0% on a 6.1 kernel. There are significant scheduler differences between the two. All crashes have MockitoAnnotations.initMocks(this); (RegisteredAidCacheTest.java(116)) in the stack for one of the threads, leading me to believe that this is being invoked at the wrong time.

Bug: 338272214
Test: atest NfcNciUnitTests

Change-Id: Iab68513d8f2f24f5f6026359d15deddbf0dca8fa
Signed-off-by: Viktor Martensson <vmartensson@google.com>